requirements.txt
crewai==0.28.8
langgraph==0.0.32
langchain==0.1.0
langchain-openai==0.0.8
openai==1.12.0
google-search-results==2.4.2
python-dotenv==1.0.0
tiktoken==0.5.2
faiss-cpu==1.7.4
chromadb==0.4.22
streamlit==1.29.0
plotly==5.17.0
pandas==2.1.4
networkx==3.2.1